move custom-mapper to separated pkg
