+++ demo/src/main/java/com/web/petros/mapper/CustomJsonMapper.java
@@ -1,4 +1,4 @@
-package com.web.petros.config;
+package com.web.petros.mapper;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
